pyproject.toml
[project]
name = "idle-clans-helper"
version = "0.1.0"
description = "Discord app to help with some information about Idle Clans game."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.2",
    "python-dotenv>=1.2.1",
]
